+++ src/allotropy/parsers/moldev_softmax_pro/softmax_pro_structure.py
@@ -81,6 +81,10 @@ class ReadType(Enum):
     SPECTRUM = "Spectrum"
     WELL_SCAN = "Well Scan"
 
+    @property
+    def is_spectrum(self) -> bool:
+        return self is ReadType.SPECTRUM
+
 
 class ExportFormat(Enum):
     TIME_FORMAT = "TimeFormat"
@@ -346,8 +350,8 @@ class PlateHeader:
     name: str
     export_version: str
     export_format: str
-    read_type: str
-    data_type: str
+    read_type: ReadType
+    data_type: DataType
     kinetic_points: int
     num_wavelengths: int
     wavelengths: list[float]
@@ -488,12 +492,10 @@ class RawData:
 class PlateRawData(RawData):
     @staticmethod
     def create(reader: CsvReader, header: PlateHeader) -> PlateRawData:
-
         columns = assert_not_none(
             reader.pop_as_series(sep="\t"),
             msg="unable to find data columns for plate block raw data.",
         )
-
         # use plate dimensions to determine how many rows of plate block to read
         dimensions = assert_not_none(
             NUM_WELLS_TO_PLATE_DIMENSIONS.get(header.num_wells),
@@ -563,12 +565,9 @@ class PlateRawData(RawData):
 
 
 @dataclass(frozen=True)
-class SpectrumRawPlateData(RawData):
-    maximum_wavelength_signal: dict[str, float | None]
-
+class SpectrumPlateRawData(RawData):
     @staticmethod
-    def create(reader: CsvReader, header: PlateHeader) -> SpectrumRawPlateData:
-
+    def create(reader: CsvReader, header: PlateHeader) -> SpectrumPlateRawData:
         columns = assert_not_none(
             reader.pop_as_series(sep="\t"),
             msg="unable to find data columns for plate block raw data.",
@@ -593,23 +592,9 @@ class SpectrumRawPlateData(RawData):
                     df_data=data.iloc[:, 2:],
                 )
             )
-        reader.pop()
         reader.drop_empty()
-        max_wavelength_signal_data = RawData.get_measurement_section(
-            reader, columns, rows
-        ).iloc[:, 2:]
-        signal_data = {
-            f"{num_to_chars(row_idx)}{col}": try_float_or_none(str(raw_value))
-            for row_idx, *row_data in max_wavelength_signal_data.itertuples()
-            for col, raw_value in zip(
-                max_wavelength_signal_data.columns, row_data, strict=True
-            )
-        }
 
-        return SpectrumRawPlateData(
-            wavelength_data=wavelength_data,
-            maximum_wavelength_signal=signal_data,
-        )
+        return SpectrumPlateRawData(wavelength_data=wavelength_data)
 
 
 @dataclass(frozen=True)
@@ -618,8 +603,6 @@ class PlateReducedData:
 
     @staticmethod
     def create(reader: CsvReader, header: PlateHeader) -> PlateReducedData:
-        if header.read_type == ReadType.SPECTRUM.value:
-            return PlateReducedData(data=[])
 
         raw_data = assert_not_none(
             reader.pop_csv_block_as_df(sep="\t", header=0),
@@ -645,7 +628,7 @@ class PlateReducedData:
 
 @dataclass(frozen=True)
 class PlateData:
-    raw_data: PlateRawData | SpectrumRawPlateData
+    raw_data: RawData
     reduced_data: PlateReducedData | None
 
     @staticmethod
@@ -653,9 +636,9 @@ class PlateData:
         reader: CsvReader,
         header: PlateHeader,
     ) -> PlateData:
-        raw_data: PlateRawData | SpectrumRawPlateData = (
-            SpectrumRawPlateData.create(reader, header)
-            if header.read_type == ReadType.SPECTRUM.value
+        raw_data: RawData = (
+            SpectrumPlateRawData.create(reader, header)
+            if header.read_type.is_spectrum
             else PlateRawData.create(reader, header)
         )
         return PlateData(
@@ -721,10 +704,12 @@ class TimeWavelengthData:
         wavelength: float,
         columns: pd.Series[str],
     ) -> TimeWavelengthData:
-        data = assert_not_none(
-            reader.pop_csv_block_as_df(sep="\t"),
-            msg="unable to find raw data from time block.",
+        raw_data = (
+            reader.pop_line_as_df(sep="\t")
+            if header.read_type.is_spectrum
+            else reader.pop_csv_block_as_df(sep="\t")
         )
+        data = assert_not_none(raw_data, msg="unable to find raw data from time block.")
         set_columns(data, columns)
 
         return TimeWavelengthData(
@@ -740,21 +725,15 @@ class TimeWavelengthData:
 class TimeRawData:
     wavelength_data: list[TimeWavelengthData]
 
-    @staticmethod
-    def create(reader: CsvReader, header: PlateHeader) -> TimeRawData:
+    @classmethod
+    def create(cls, reader: CsvReader, header: PlateHeader) -> TimeRawData:
         columns = assert_not_none(
             reader.pop_as_series(sep="\t"),
             msg="unable to find data columns for time block raw data.",
         )
-
         return TimeRawData(
             wavelength_data=[
-                TimeWavelengthData.create(
-                    reader,
-                    header,
-                    wavelength,
-                    columns,
-                )
+                TimeWavelengthData.create(reader, header, wavelength, columns)
                 for wavelength in header.wavelengths
             ]
         )
@@ -799,11 +778,11 @@ class TimeData:
         reader: CsvReader,
         header: PlateHeader,
     ) -> TimeData:
-        raw_data = None
         reduced_data = None
+        raw_data: TimeRawData | None = None
 
         # Read raw data if data_type is RAW or BOTH
-        if header.data_type in (DataType.RAW.value, DataType.BOTH.value):
+        if header.data_type in (DataType.RAW, DataType.BOTH):
             raw_data = TimeRawData.create(reader, header)
         # For REDUCED only, create synthetic raw data with error message
         else:
@@ -821,8 +800,8 @@ class TimeData:
     @staticmethod
     def _create_synthetic_raw_data(header: PlateHeader) -> TimeRawData:
         """Create synthetic raw data with error messages when only reduced data is available."""
-        synthetic_wavelength_data = []
 
+        synthetic_wavelength_data = []
         for wavelength in header.wavelengths:
             # For each position in the plate, create a data element with an error document
             num_cols = header.num_columns
@@ -905,7 +884,7 @@ class PlateBlock(ABC, Block):
 
     @property
     def measurement_type(self) -> MeasurementType:
-        if self.header.read_type == ReadType.SPECTRUM.value:
+        if self.header.read_type is ReadType.SPECTRUM:
             if self.header.concept == "absorbance":
                 return MeasurementType.ULTRAVIOLET_ABSORBANCE_CUBE_SPECTRUM
             elif self.header.concept == "fluorescence":
@@ -925,7 +904,7 @@ class PlateBlock(ABC, Block):
         return read_mode_to_measurement_type[self.header.read_mode]
 
     @classmethod
-    def parse_header(cls, header: pd.Series[str]) -> PlateHeader:
+    def parse_header(cls, _: pd.Series[str]) -> PlateHeader:
         raise NotImplementedError
 
     @classmethod
@@ -935,28 +914,30 @@ class PlateBlock(ABC, Block):
             raise AllotropeConversionError(msg)
 
     @classmethod
-    def check_read_type(cls, read_type: str) -> None:
-        if read_type not in (
+    def check_read_type(cls, read_type_raw: str) -> ReadType:
+        if read_type_raw not in (
             ReadType.ENDPOINT.value,
             ReadType.KINETIC.value,
             ReadType.SPECTRUM.value,
         ):
-            msg = f"Only Endpoint, Spectrum or Kinetic measurements can be processed at this time, got: {read_type}"
+            msg = f"Only Endpoint, Spectrum or Kinetic measurements can be processed at this time, got: {read_type_raw}"
             raise AllotropeConversionError(msg)
+        return ReadType(read_type_raw)
 
     @classmethod
-    def get_read_mode(cls, read_type: str, read_mode_raw: str) -> str:
-        return f"{'Kinetic ' if read_type == ReadType.KINETIC.value else ''}{read_mode_raw}"
+    def get_read_mode(cls, read_type: ReadType, read_mode_raw: str) -> str:
+        return f"{'Kinetic ' if read_type is ReadType.KINETIC else ''}{read_mode_raw}"
 
     @classmethod
-    def check_data_type(cls, data_type: str) -> None:
-        if data_type not in (
+    def check_data_type(cls, data_type_raw: str) -> DataType:
+        if data_type_raw not in (
             DataType.RAW.value,
             DataType.BOTH.value,
             DataType.REDUCED.value,
         ):
-            msg = f"Unexpected data type: {data_type}, supported values are RAW, REDUCED, or BOTH."
+            msg = f"Unexpected data type: {data_type_raw}, supported values are RAW, REDUCED, or BOTH."
             raise AllotropeConversionError(msg)
+        return DataType(data_type_raw)
 
     @classmethod
     def check_num_wavelengths(
@@ -1018,10 +999,10 @@ class FluorescencePlateBlock(PlateBlock):
             name,
             export_version,
             export_format,
-            read_type,
+            read_type_raw,
             read_mode_raw,
             raw_scan_position,
-            data_type,
+            data_type_raw,
             _,  # Pre-read, always FALSE
             kinetic_points_raw,
             read_time,  # read_time_or_scan_pattern
@@ -1048,9 +1029,9 @@ class FluorescencePlateBlock(PlateBlock):
         ] = header[:31]
 
         cls.check_export_version(export_version)
-        cls.check_read_type(read_type)
-        cls.check_data_type(data_type)
-        if read_type == ReadType.SPECTRUM.value:
+        read_type = cls.check_read_type(read_type_raw)
+        data_type = cls.check_data_type(data_type_raw)
+        if read_type is ReadType.SPECTRUM:
             msg = (
                 "Spectrum read type is not currently supported for fluorescence plates."
             )
@@ -1139,9 +1120,9 @@ class LuminescencePlateBlock(PlateBlock):
             name,
             export_version,
             export_format,
-            read_type,
+            read_type_raw,
             read_mode_raw,
-            data_type,
+            data_type_raw,
             _,  # Pre-read, always FALSE
             kinetic_points_raw,
             read_time,  # read_time_or_scan_pattern
@@ -1168,9 +1149,9 @@ class LuminescencePlateBlock(PlateBlock):
         ] = header[:30]
 
         cls.check_export_version(export_version)
-        cls.check_read_type(read_type)
-        cls.check_data_type(data_type)
-        if read_type == ReadType.SPECTRUM.value:
+        read_type = cls.check_read_type(read_type_raw)
+        data_type = cls.check_data_type(data_type_raw)
+        if read_type is ReadType.SPECTRUM:
             msg = (
                 "Spectrum read type is not currently supported for luminescence plates."
             )
@@ -1217,9 +1198,9 @@ class AbsorbancePlateBlock(PlateBlock):
             name,
             export_version,
             export_format,
-            read_type,
+            read_type_raw,
             read_mode_raw,
-            data_type,
+            data_type_raw,
             _,  # Pre-read, always FALSE
             kinetic_points_raw,
             read_time,  # read_time_or_scan_pattern
@@ -1237,15 +1218,15 @@ class AbsorbancePlateBlock(PlateBlock):
         ] = header[:21]
 
         cls.check_export_version(export_version)
-        cls.check_read_type(read_type)
-        cls.check_data_type(data_type)
+        read_type = cls.check_read_type(read_type_raw)
+        data_type = cls.check_data_type(data_type_raw)
 
         num_wavelengths = cls.get_num_wavelengths(num_wavelengths_raw)
-        if read_type == ReadType.SPECTRUM.value:
+        if read_type is ReadType.SPECTRUM:
             num_wavelengths = try_int(kinetic_points_raw, "kinetic_points")
 
         wavelengths = cls.get_wavelengths(wavelengths_str)
-        if read_type == ReadType.SPECTRUM.value:
+        if read_type is ReadType.SPECTRUM:
             wavelengths = cls.get_wavelengths_from_start_end_step(
                 start_wavelength, end_wavelength, wavelength_step
             )
@@ -1361,9 +1342,6 @@ class StructureData:
                         # experiment, there is no way at the moment to link the group data with a plate.
                         continue
                     plate_block = block_list.plate_blocks[group_data_element.plate]
-                    is_spectrum = (
-                        plate_block.header.read_type == ReadType.SPECTRUM.value
-                    )
                     processed_positions = set()
 
                     for data_element in plate_block.iter_data_elements(
@@ -1374,11 +1352,11 @@ class StructureData:
 
                         # For spectrum measurements, only add errors to the first DataElement per well
                         if (
-                            not is_spectrum
+                            not plate_block.header.read_type.is_spectrum
                             or data_element.position not in processed_positions
                         ):
                             data_element.error_document += group_data_element.errors
-                            if is_spectrum:
+                            if plate_block.header.read_type.is_spectrum:
                                 processed_positions.add(data_element.position)
 
                         data_element.custom_info.update(group_data_element.custom_info)