+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this packages will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.99] - 2025-07-15
+
+### Added
+
+- Molecular Devices SoftMax Pro - add support for spectral TimeFormat measurements (#1026)
+
+### Fixed
+
+- Molecular Devices SoftMax Pro - fix duplicate data source documents in spectrum measurements (#1033)
+- Molecular Devices SoftMax Pro - move reduced data from custom info to calculated data document (#1032)
+
 ## [0.1.98] - 2025-07-08
 
 ### Added